requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers